// src/core/types.d.ts
export type stringOrExpr = string | {
    expr: string;
};
// ...
export interface BaseControlType {
    /** Sets the `class` attribute on the component or control.
     *
     * - Can be applied at the component level (`ion-item`):
     *
     * ```typescript
     * component: { class: 'car-sport pretty' },
     * ```
     * ```html
     * <ion-item class="car-sport pretty">
     *   <ion-label></ion-label>
     * </ion-item>
     * ```
     * - ...or at the control level (`ion-input`, `ion-toggle`, `ion-button`, `ion-datetime`, `ion-checkbox`, `ion-icon`, `ion-label`):
     *
     * ```typescript
     * component: {
     *   label: { class: 'car-sport pretty' }
     * }
     * ```
     * ```html
     * <ion-item>
     *   <ion-label class="car-sport pretty"></ion-label>
     * </ion-item>
     * ```
     */
    class?: string | undefined;
    /** Adds and removes CSS classes on an HTML element.
     *
     * - Can be applied at the component level (`ion-item`):
     *
     * ```typescript
     * component: { ngClass: `{'error': invalid }` },
     * ```
     * ```html
     * <ion-item [ngClass]="{'error': invalid }">
     *   <ion-input></ion-input>
     * </ion-item>
     * ```
     * - ...or at the control level (`ion-input`, `ion-toggle`, `ion-button`, `ion-datetime`, `ion-checkbox`, `ion-icon`, `ion-label`):
     *
     * ```typescript
     * component: {
     *   input: { ngClass: `{'error': invalid }` }
     * }
     * ```
     * ```html
     * <ion-item>
     *   <ion-input [ngClass]="{'error': invalid }"></ion-input>
     * </ion-item>
     * ```
     */
    ngClass?: string | undefined;
    /** Sets the `style` attribute on the component or control.
     *
     * - Can be applied at the component level (`ion-item`):
     *
     * ```typescript
     * component: { style: 'margin-top: 10px; font-size: 14px;' },
     * ```
     * ```html
     * <ion-item style="margin-top: 10px; font-size: 14px;">
     *   <ion-label></ion-label>
     * </ion-item>
     * ```
     * - ...or at the control level (`ion-input`, `ion-toggle`, `ion-button`, `ion-datetime`, `ion-checkbox`, `ion-icon`, `ion-label`):
     *
     * ```typescript
     * component: {
     *   label: { style: 'margin-top: 10px; font-size: 14px;' }
     * }
     * ```
     * ```html
     * <ion-item>
     *   <ion-label style="margin-top: 10px; font-size: 14px;"></ion-label>
     * </ion-item>
     * ```
     */
    style?: string | undefined;
    /** Adds and removes CSS styles on an HTML element.
     *
     * - Can be applied at the component level (`ion-item`):
     *
     * ```typescript
     * component: { ngStyle: `{'error': invalid }` },
     * ```
     * ```html
     * <ion-item [ngStyle]="{'error': invalid }">
     *   <ion-input></ion-input>
     * </ion-item>
     * ```
     * - ...or at the control level (`ion-input`, `ion-toggle`, `ion-button`, `ion-datetime`, `ion-checkbox`, `ion-icon`, `ion-label`):
     *
     * ```typescript
     * component: {
     *   input: { ngStyle: `{'error': invalid }` }
     * }
     * ```
     * ```html
     * <ion-item>
     *   <ion-input [ngStyle]="{'error': invalid }"></ion-input>
     * </ion-item>
     * ```
     */
    ngStyle?: string | undefined;
    /** Adds and removes CSS styles on an HTML element.
     *
     * - Can be applied at the row level (`ion-row`):
     *
     * ```typescript
     * row: { expanded: `row.numVehiculos === 1 || ui.expand.caracteristicas` },
     * ```
     * ```html
     * <ion-row [expanded]="row.numVehiculos === 1 || ui.expand.caracteristicas"></ion-row>
     * ```
     */
    expanded?: string | undefined;
}
// ...
export interface FieldType {
    /** Field name. Used with `formControlName` to bind the control in reactive forms. */
    name?: string;
    /** Declares the value type of the component for the front-end.
     *
     * ```typescript
     * value: 'boolean'
     * ```
     * To set a default value:
     * ```typescript
     * value: { type: 'boolean', default: true }
     * ```
     *
     * ```typescript
     * const defaultValue = field.value.default;
     * const control = new FormControl(defaultValue, validators)
     * ```
     */
    value?: ValueType | {
        type: ValueType;
        default?: any;
    } | undefined;
    column?: FieldColumnType;
    component?: FieldComponentType;
}
export interface FieldTemplateType extends FieldType {
    /** Field name to use when referencing a template without an explicit `name`. */
    template: string;
}
export interface DynamicFieldsRow {
    cols: FieldType[];
    frm?: FormGroup;
    host?: any;
}
export interface DynamicRowType {
    title?: stringOrExpr;
    row?: BaseControlType;
    cols?: FieldType[];
    permission?: string | boolean;
    fields?: any;
}
export interface DynamicFieldsGrid {
    grid?: BaseControlType;
    rows?: DynamicRowType[];
    frm?: FormGroup;
    host?: any;
}